:truck: (GameEngineKit): Move extensions to Utils
