+++ Modules/GameEngineKit/Sources/UI/Views/PlayButton.swift
@@ -5,16 +5,6 @@
 import DesignKit
 import SwiftUI
 
-extension Shape {
-    func fill<Fill: ShapeStyle, Stroke: ShapeStyle>(
-        _ fillStyle: Fill, strokeBorder strokeStyle: Stroke, lineWidth: CGFloat = 1
-    ) -> some View {
-        self
-            .stroke(strokeStyle, lineWidth: lineWidth)
-            .background(self.fill(fillStyle))
-    }
-}
-
 struct Play_ButtonStyle: ButtonStyle {
     var progress: CGFloat
